Temp

+++ backend/src/main/java/za/ac/cput/tekkiestorecapstone/domain/OrderItem.java
@@ -13,6 +13,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.math.BigDecimal;
 
 // Represents an item in an order.
 @Entity
@@ -34,8 +35,8 @@ public class OrderItem {
     private String imageUrl;
 
     private int quantity;
-    private double unitPrice;
-    private double subTotal;
+    private BigDecimal unitPrice;
+    private BigDecimal subTotal;
 
     // Default constructor
     protected OrderItem() {
@@ -92,15 +93,15 @@ public class OrderItem {
         return quantity;
     }
 
-    public double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
 
-    public double getSubTotal() {
+    public BigDecimal getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(double subTotal) {
+    public void setSubTotal(BigDecimal subTotal) {
         this.subTotal = subTotal;
     }
 
@@ -131,8 +132,8 @@ public class OrderItem {
         private String size;
         private String imageUrl;
         private int quantity;
-        private double unitPrice;
-        private double subTotal;
+        private BigDecimal unitPrice;
+        private BigDecimal subTotal;
 
         public Builder setOrderItemId(String orderItemId) {
             this.orderItemId = orderItemId;
@@ -174,12 +175,12 @@ public class OrderItem {
             return this;
         }
 
-        public Builder setUnitPrice(double unitPrice) {
+        public Builder setUnitPrice(BigDecimal unitPrice) {
             this.unitPrice = unitPrice;
             return this;
         }
 
-        public Builder setSubTotal(double subTotal) {
+        public Builder setSubTotal(BigDecimal subTotal) {
             this.subTotal = subTotal;
             return this;
         }
